refactor: move all goal in review files to one util, instead of multiples per component
